% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_pl.R
\name{ternary_pl}
\alias{ternary_pl}
\title{Data structure with the  projected and boundary by node and class.}
\usage{
ternary_pl(dat, dx, dy)
}
\arguments{
\item{dat}{data.frame}

\item{dx}{first direction included in id}

\item{dy}{second direction included in id}
}
\value{
ternary plot
}
\description{
Data structure with the  projected and boundary by node and class.
}
\examples{
#crab data set with all the observations used as training
pprf.crab <- PPforest::PPforest(data = crab, std =TRUE, class = "Type",
 size.tr = 1, m = 100, size.p = .5, PPmethod = 'LDA')
 st <- ternary_str(pprf.crab, id = c(1, 2, 3), sp = 3, dx = 1, dy = 2)
 ternary_pl(st, dx =1, dy = 2)
}
